transient-bib-mode inherits from special-mode
Doing this allows for buffers that use this mode directly be
read-only, which is one of the key features of this package. (One
cannot accidentally and easily alter their bibliography database.)
In addition, inheriting from special-mode overrides many of the normal
global keymap bindings, such as (self-insert-command), but leaves
mark-moving commands well-defined.