First implementation of adding @Article entry to file
We start by getting the current buffer (the main bib file) and making
an indirect buffer that is attached to the main buffer. From there,
the main bib file is saved at the beginning of (save-current-buffer...
The body of the form is evaluated, and the last thing the
save-current-buffer form SHOULD do is switch us back to the main
bibliography file.