Start work on macro to handle opening entries for use
I hope that this macro will be able to accept the necessary arguments
to open an arbitrary entry in a temporary buffer for editing, and when
a keybinding is pressed, the temporary buffer is closed, with the main
bibliography file being edited in-place and saved.

If the entry already exists AND the point (cursor) is inside the
entry, then only the entry that surrounds the point should be opened
in the temporary buffer. If the entry exists, but the point is NOT
inside an entry, then a search query should happen to narrow to an
entry. If the entry does NOT exist at all, then open a new buffer and
ask for which kind of entry to insert.